add 16 bit integer ranges to allow for larger ints in wpm

+++ src/CrossPointSettings.h
@@ -62,7 +62,7 @@ class CrossPointSettings {
   uint8_t extraParagraphSpacing = 1;
   uint8_t textAntiAliasing = 1;
   // Reading speed (words per minute) for time-left estimate
-  uint8_t readingSpeedWpm = 200;
+  uint16_t readingSpeedWpm = 200;
   // Toggle to show time remaining in the current chapter
   uint8_t showTimeLeftInChapter = 0;
   // Duration of the power button press
